+++ shared/r2-shared-swift/Toolkit/Extensions/Optional.swift
@@ -1,12 +1,7 @@
 //
-//  Optional.swift
-//  r2-shared-swift
-//
-//  Created by Mickaël Menu on 12/07/2020.
-//
 //  Copyright 2020 Readium Foundation. All rights reserved.
-//  Use of this source code is governed by a BSD-style license which is detailed
-//  in the LICENSE file present in the project repository where this source code is maintained.
+//  Use of this source code is governed by the BSD-style license
+//  available in the top-level LICENSE file of the project.
 //
 
 import Foundation
@@ -21,6 +16,17 @@ extension Optional {
         case .none:
             throw error()
         }
+    }
+
+    /// Returns `nil` if the value doesn't pass the given `condition`.
+    public func takeIf(_ condition: (Wrapped) -> Bool) -> Self {
+        guard
+            case .some(let value) = self,
+            condition(value)
+        else {
+            return nil
+        }
+        return value
     }
     
 }