+++ shared/r2-shared-swift/Toolkit/Extensions/URL.swift
@@ -1,12 +1,7 @@
 //
-//  URL.swift
-//  r2-shared-swift
-//
-//  Created by Mickaël Menu on 12/07/2020.
-//
 //  Copyright 2020 Readium Foundation. All rights reserved.
-//  Use of this source code is governed by a BSD-style license which is detailed
-//  in the LICENSE file present in the project repository where this source code is maintained.
+//  Use of this source code is governed by the BSD-style license
+//  available in the top-level LICENSE file of the project.
 //
 
 import Foundation
@@ -18,14 +13,14 @@ extension URL: Loggable {
     public var isHTTP: Bool {
         ["http", "https"].contains(scheme?.lowercased())
     }
-    
+
     /// Returns whether the given `url` is `self` or one of its descendants.
     public func isParentOf(_ url: URL) -> Bool {
         let standardizedSelf = standardizedFileURL.path
         let other = url.standardizedFileURL.path
         return standardizedSelf == other || other.hasPrefix(standardizedSelf + "/")
     }
-    
+
     /// Computes the MD5 hash of the file, if the URL is a file URL.
     /// Source: https://stackoverflow.com/a/42935601/1474476
     public func md5() -> String? {