+++ shared/r2-shared-swiftTests/Fetcher/ArchiveFetcherTests.swift
@@ -69,7 +69,7 @@ class ArchiveFetcherTests: XCTestCase {
         let resource = fetcher.get(Link(href: "/unknown"))
         let result = resource.read()
         XCTAssertThrowsError(try result.get()) { error in
-            XCTAssertEqual(ResourceError.notFound, error as? ResourceError)
+            XCTAssertEqual(ResourceError.notFound(nil), error as? ResourceError)
         }
     }
     
@@ -82,7 +82,7 @@ class ArchiveFetcherTests: XCTestCase {
         let resource = fetcher.get(Link(href: "/unknown"))
         let result = resource.length
         XCTAssertThrowsError(try result.get()) { error in
-            XCTAssertEqual(ResourceError.notFound, error as? ResourceError)
+            XCTAssertEqual(ResourceError.notFound(nil), error as? ResourceError)
         }
     }